\name{owo_whats_this}
\alias{owo_whats_this}
\title{ohayo i pwesent beta-nyaatenin uwu}
\usage{
owo_whats_this()
}
\description{
ÒwÓ what's this? bwomeodomain made a package? spicy wicy~

Ok in all seriousness, this package is a combination of my previous (now deleted) packages, both pertaining to molecular biology techniques and geometric morphometrics. pnplandmark placed 8 fiducials around the border of the posterior neuropore, drawing from two open curves and one designated point of fusion. qPCR included formulae and analysis for real-time (quantitative) polymerase chain reaction (qPCR) setup and execution, including RNA extraction and reverse transcription. betanyaatenin will include functions found in those packages as well as those I draft along the way.

This package was initialized on 2021-01-07 by https://github.com/bromeodomain (mcmahon) and is maintained by the same individual. I hope nobody has to read this as there are undoubtedly better packages out there.
}
