\name{secant_length}
\alias{secant_length}
\title{
Determine Direct Length Between Two Points
}
\description{
Determines direct length between two points, not accounting for tissue curvature.
}
\usage{
secant_length(F1, F2)
}
\arguments{
  \item{F1}{one endpoint}
  \item{F2}{another endpoint}
}
\details{
This is literally just the Pythagorean Theorem in 3 dimensions. Nothing to see here. Output is just a double. That's it. I don't know what you expected, but I suggest you lower your expectations. I am short. I cannot reach them. Send help.
}
\examples{
secant_length(F1, F2)
}
